# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ManagementSystem API",
        "description": "API documentation for ManagementSystem",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/ManagementSystem/public",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users with pagination",
                "description": "Returns paginated list of all users with their projects and roles. Requires admin or owner role.",
                "operationId": "eaedf6aa84fde38a08bf5d87308169f7",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Users retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/UserResource"
                                                    }
                                                },
                                                "links": {
                                                    "properties": {
                                                        "first": {
                                                            "type": "string",
                                                            "example": "http://example.com/api/users?page=1"
                                                        },
                                                        "last": {
                                                            "type": "string",
                                                            "example": "http://example.com/api/users?page=5"
                                                        },
                                                        "prev": {
                                                            "type": "string",
                                                            "example": null,
                                                            "nullable": true
                                                        },
                                                        "next": {
                                                            "type": "string",
                                                            "example": "http://example.com/api/users?page=2",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "meta": {
                                                    "properties": {
                                                        "current_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "from": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "last_page": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "path": {
                                                            "type": "string",
                                                            "example": "http://example.com/api/users"
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "to": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 50
                                                        },
                                                        "links": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "url": {
                                                                        "type": "string",
                                                                        "example": "http://example.com/api/users?page=1",
                                                                        "nullable": true
                                                                    },
                                                                    "label": {
                                                                        "type": "string",
                                                                        "example": "1"
                                                                    },
                                                                    "active": {
                                                                        "type": "boolean",
                                                                        "example": true
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized - Admin or Owner role required"
                    },
                    "429": {
                        "description": "Too Many Requests - Throttled"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/role/{roleName}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get users by role",
                "description": "Returns all users who have the specified role. Requires admin or owner role.",
                "operationId": "a5d06527e3e4e57120ebe69075d92d9b",
                "parameters": [
                    {
                        "name": "roleName",
                        "in": "path",
                        "description": "Role name to filter by",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/RolesEnum"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Users retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized - Admin or Owner role required"
                    },
                    "422": {
                        "description": "Invalid role name"
                    },
                    "429": {
                        "description": "Too Many Requests - Throttled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/project/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get users assigned to a project",
                "description": "Returns all users assigned to the specified project. Requires admin or owner role.",
                "operationId": "91f043ff0a8cd78df560cf96090984de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Users retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized - Admin or Owner role required"
                    },
                    "404": {
                        "description": "Project not found"
                    },
                    "429": {
                        "description": "Too Many Requests - Throttled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Change user roles",
                "description": "Assign new roles to a user. Admin cannot modify other admins or owners.",
                "operationId": "08a7e2b53f7485e0b976bc7d57221cb5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "names"
                                ],
                                "properties": {
                                    "names": {
                                        "description": "Array of role names",
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/RolesEnum"
                                        },
                                        "example": [
                                            "backend",
                                            "frontend"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User role updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User role updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized - Admin cannot modify other admins/owners"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation error - Invalid role names"
                    },
                    "429": {
                        "description": "Too Many Requests - Throttled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "RolesEnum": {
                "title": "RolesEnum",
                "description": "Available user roles in the system",
                "type": "string",
                "example": "admin",
                "enum": [
                    "owner",
                    "admin",
                    "backend",
                    "frontend",
                    "ui"
                ]
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}